Implemented a class to evaluate algorithm, i.e. to compute competitve ratios by using Monte-Carlo simulations

+++ src/graph/Matching.java
@@ -26,7 +26,7 @@ public class Matching {
         return size; 
     }
 
-    public double getWeight() {
+    public int getWeight() {
         return weight; 
     }
